refactor(server-nestjs): reuse VAULT_PROVISION_TIMEOUT, rename e2e-timeout to constants
