+++ apps/server-nestjs/test/constants.ts
@@ -3,9 +3,5 @@
 export const SONARQUBE_PROJECT_TIMEOUT = 30_000 // provision + delete a SonarQube project/user
 export const KEYCLOAK_GROUP_SYNC_TIMEOUT = 60_000 // reconcile Keycloak groups/roles
 export const EXTERNAL_SYNC_TIMEOUT = 72_000 // sync GitLab groups/members, teardown Nexus/registry
-// A project reconcile that also exercises the orphan purge runs the gitlab plugin twice
-// (create + purge) on a shared integration environment; the 72s EXTERNAL_SYNC_TIMEOUT is
-// not enough for that double pass.
-export const GITLAB_PURGE_RECONCILE_TIMEOUT = 150_000
 export const ARGOCD_RECONCILE_TIMEOUT = 144_000 // ArgoCD commit + sync to git
 export const VAULT_PROVISION_TIMEOUT = 180_000 // provision Vault mounts/policies/approles, zone secrets
